cast to 64bit integer

+++ csrc/cuda/utils.cuh
@@ -9,19 +9,19 @@
 __device__ __inline__ at::Half __shfl_up_sync(const unsigned mask,
                                               const at::Half var,
                                               const unsigned int delta) {
-  return __shfl_up_sync(mask, var.operator __half(), delta);
+  return __shfl_up_sync((unsigned long long)mask, var.operator __half(), delta);
 }
 
 __device__ __inline__ at::Half __shfl_down_sync(const unsigned mask,
                                                 const at::Half var,
                                                 const unsigned int delta) {
-  return __shfl_down_sync(mask, var.operator __half(), delta);
+  return __shfl_down_sync((unsigned long long)mask, var.operator __half(), delta);
 }
 
 __device__ __inline__ at::Half __shfl_sync(const unsigned mask,
                                            const at::Half var,
                                            const int delta) {
-  return __shfl_sync(mask, var.operator __half(), delta);
+  return __shfl_sync((unsigned long long)mask, var.operator __half(), delta);
 }
 
 __device__ __inline__ at::Half __shfl_up(const at::Half var,
